+++ CHANGELOG.md
@@ -20,3 +20,4 @@ v0.4
 - fixed line_count not resetting after a game_over
 - show battery indicator
 - sticky direction are now working
+- simple version check